Moved functions to util, might have a huge refactor job coming in soon, just incase it doesnt workout gonna keep it on a separate branch, currently cant delete delimiters for some reason
