+++ config/webpack.dev.config.js
@@ -15,34 +15,6 @@ module.exports = {
         test: /\.js$/,
         exclude: /node_modules/,
         use: ['babel-loader']
-      },
-      {
-        test: /\.css$/,
-        exclude: /node_modules/,
-        use: ['style-loader', 'css-loader']
-      },
-      {
-        test: /\.(jpe?g|png|gif|ico)$/i,
-        use: ['file-loader']
-      },
-      {
-        test: /\.sass$/,
-        use: [
-          'style-loader',
-          'css-loader',
-          'sass-loader'
-        ]
-      },
-      {
-        test: /\.(jpg|png|gif|svg|pdf|ico)$/,
-        use: [
-          {
-            loader: 'file-loader',
-            options: {
-              outputPath: 'images'
-            }
-          }
-        ]
       }
     ]
   },
